Move json-rpc-server.ts
